wraps nodemailer sendMail function in a promise
